Create Tic-Tac-Toe.for
This is a fortran version of the game TIc-Tac-Toe. If there are any compile errors try to check the position for each error. The most of the times try to use the space in order to correct them. In this program there is used the command goto and the position that is used each time may lead to errors if you do not put this number to the correct place. Thanks a lot!